Crate alloy_signer_local

alloy-signer-local

Local signer implementations:

K256 private key

Mnemonic phrase

YubiHSM2

Features

keystore: enables Ethereum keystore functionality on the PrivateKeySigner type.

mnemonic: enables BIP-39 mnemonic functionality for building PrivateKeySigners.

yubihsm: enables LocalSigners with YubiHSM2 support.

Re-exports

pub use yubihsm;

yubihsm

pub use coins_bip39;

mnemonic

Modules

error 🔒

mnemonic 🔒 mnemonic

Specific helper functions for creating/loading a mnemonic private key following BIP-39 specifications.

private_key 🔒

[k256] signer implementation.

yubi 🔒 yubihsm

YubiHSM2 signer implementation.

Structs

LocalSigner

An Ethereum private-public key pair which can be used for signing messages.

MnemonicBuildermnemonic

Represents a structure that can resolve into a PrivateKeySigner.

Enums

LocalSignerError

Error thrown by LocalSigner.

Type Aliases

PrivateKeySigner

A signer instantiated with a locally stored private key.

YubiSigneryubihsm

A signer instantiated with a YubiHSM.