Crate alloy_signer_local

alloy-signer-local

Local signer implementations:

K256 private key

Mnemonic phrase

YubiHSM2

Features

keystore: enables Ethereum keystore functionality on the PrivateKeySigner type.

mnemonic: enables BIP-39 mnemonic functionality for building PrivateKeySigners.

yubihsm: enables LocalSigners with YubiHSM2 support.

Structs

LocalSigner

An Ethereum private-public key pair which can be used for signing messages.

Enums

LocalSignerError

Error thrown by LocalSigner.

Type Aliases

PrivateKeySigner

A signer instantiated with a locally stored private key.